Fix warnings in glslang
glslang/MachineIndependent/ParseHelper.cpp:1510:13: error: 712 enumeration values not handled in switch: 'EOpNull', 'EOpSequence', 'EOpLinkerObjects'... [-Werror,-Wswitch]
    switch (callNode.getOp()) {
            ^
1 error generated.

glslang/SPIRV/SpvBuilder.cpp:643:68: error: comparison of integers of different signs: 'unsigned int' and 'int' [-Werror,-Wsign-compare]
        return typeClass == typeOp && instr.getImmediateOperand(0) == width;
                                      ~~~~~~~~~~~~~~~~~~~~~~~~~~~~ ^  ~~~~~
1 error generated.

+++ SPIRV/SpvBuilder.cpp
@@ -640,7 +640,7 @@ bool Builder::containsType(Id typeId, spv::Op typeOp, int width) const
     {
     case OpTypeInt:
     case OpTypeFloat:
-        return typeClass == typeOp && instr.getImmediateOperand(0) == width;
+        return typeClass == typeOp && instr.getImmediateOperand(0) == (unsigned int)width;
     case OpTypeStruct:
         for (int m = 0; m < instr.getNumOperands(); ++m) {
             if (containsType(instr.getIdOperand(m), typeOp, width))
